Refactored/reorganized Modal into one component with interchangeable sub-components. Working toward greater reusability and readability with the modal, and reducing duplicate code. Ultimate goal is to update the way threads are stored in the database (instead of storing text as part of threads, store text as a reply to the new thread) and then update Modal components accordingly. Still some code overlap between the thread form and the reply form.
